+++ main-app/src/Components/MainSites/Footer/Cookies.js
