+++ main-app/src/Components/MainSites/Footer/Privacy.js
