+++ main-app/src/Components/MainSites/MainMenuHeader/Home.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Tile from "../Tile/Tile";
+import Tile from "../../Tile/Tile";
 
 //import "../../Styles/MainSites/Grid_home.css"
