+++ main-app/src/Components/MainSites/TilesHierarchy/Loci.js
@@ -1,8 +1,8 @@
 
 import React from 'react';
-import Tile from "../Tile/Tile";
+import Tile from "../../Tile/Tile";
 
-import "../../Styles/MainSites/Loci.css"
+import "../../../Styles/MainSites/Loci.css"
 
 const Loci = () => {
     return (
