+++ main-app/src/Components/MainSites/TilesHierarchy/Major.js
@@ -1,8 +1,8 @@
 
 import React from 'react';
-import Tile from "../Tile/Tile";
+import Tile from "../../Tile/Tile";
 
-import "../../Styles/MainSites/Major.css"
+import "../../../Styles/MainSites/Major.css"
 
 const Major = () => {
     return (
