+++ main-app/src/Components/MainSites/TilesHierarchy/Mind-Palace.js
@@ -1,8 +1,8 @@
 
 import React from 'react';
-import Tile from "../Tile/Tile";
+import Tile from "../../Tile/Tile";
 
-import "../../Styles/MainSites/MindPalace.css"
+import "../../../Styles/MainSites/MindPalace.css"
 
 const MindPalace = () => {
     return (
