+++ main-app/src/Components/MainSites/TilesHierarchy/Names.js
@@ -1,8 +1,8 @@
 
 import React from 'react';
-import Tile from "../Tile/Tile";
+import Tile from "../../Tile/Tile";
 
-import "../../Styles/MainSites/Names.css"
+import "../../../Styles/MainSites/Names.css"
 
 const Names = () => {
     return (
