+++ main-app/src/Components/MainSites/TilesHierarchy/Pegs.js
@@ -1,8 +1,8 @@
 
 import React from 'react';
-import Tile from "../Tile/Tile";
+import Tile from "../../Tile/Tile";
 
-import "../../Styles/MainSites/Pegs.css"
+import "../../../Styles/MainSites/Pegs.css"
 
 const Pegs = () => {
     return (
