+++ main-app/src/Components/MainSites/TilesHierarchy/Vocab.js
@@ -1,8 +1,8 @@
 
 import React from 'react';
-import Tile from "../Tile/Tile";
+import Tile from "../../Tile/Tile";
 
-import "../../Styles/MainSites/Vocab.css"
+import "../../../Styles/MainSites/Vocab.css"
 
 const Vocab = () => {
     return (
