+++ main-app/src/Components/MnemoMaxTrainer/MnemoClock.js
@@ -3,7 +3,7 @@ import React, {useState, useEffect, useContext} from "react";
 
 // Import contexts
 import {AppState
-} from "../MainSites/MainMenu/MnemoContexts";
+} from "../MainSites/MainMenuHeader/MnemoContexts";
 import App from "../../App";
 
 const MnemoClock = () => {
