+++ main-app/src/Components/MnemoMaxTrainer/MnemoContent.js
@@ -4,7 +4,7 @@ import React, {useCallback, useContext, useEffect} from "react";
 
 import {
     AppState
-} from "../MainSites/MainMenu/MnemoContexts";
+} from "../MainSites/MainMenuHeader/MnemoContexts";
 
 import MnemoTest from "./MnemoTest";
 import MnemoResult from "./MnemoResult";
