+++ main-app/src/Components/MnemoMaxTrainer/MnemoResult.js
@@ -1,7 +1,7 @@
 import React, {useContext, useEffect, useState} from "react";
 import {
     AppState
-} from "../MainSites/MainMenu/MnemoContexts";
+} from "../MainSites/MainMenuHeader/MnemoContexts";
 
 
 const MnemoResult = () => {
