+++ main-app/src/Components/MnemoMaxTrainer/MnemoSettings.js
@@ -1,6 +1,6 @@
 import React, {useContext, useState} from "react";
 import App from "../../App";
-import {AppState} from "../MainSites/MainMenu/MnemoContexts";
+import {AppState} from "../MainSites/MainMenuHeader/MnemoContexts";
 import MnemoButton from "./MnemoButton";
