+++ main-app/src/Components/MnemoMaxTrainer/MnemoTrain.js
@@ -2,7 +2,7 @@ import React, {useContext, useEffect, useState} from "react";
 import {
 AppState
 
-} from "../MainSites/MainMenu/MnemoContexts";
+} from "../MainSites/MainMenuHeader/MnemoContexts";
